# docs/openapi/dbcheck-web.yaml
openapi: 3.0.3
info:
  title: db-check Web Report API
  version: 0.1.0
  description: |
    Web service for generating db-check reports from uploaded collector ZIP packages.
servers:
  - url: /
paths:
  /api/reports/generate:
    post:
      summary: Generate reports from uploaded ZIP packages
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - zips
              properties:
                zips:
                  type: array
                  description: One or more collector ZIP packages.
                  items:
                    type: string
                    format: binary
                awrs:
                  type: array
                  description: |
                    Optional Oracle AWR HTML files.
                    Either provide `awrs` with the same count as `zips` (paired by index),
                    or provide per-item fields `awr_<index>` (1-based, e.g. `awr_1`).
                  items:
                    type: string
                    format: binary
      responses:
        "200":
          description: Task created and processing started (queued on a single worker).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateResponse"
        "400":
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Upload too large.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
  schemas:
    GenerateResponse:
      type: object
      required:
        - task_id
        - status
        - total
        - ws_url
      properties:
        task_id:
          type: string
        status:
          type: string
          enum: ["processing"]
        total:
          type: integer
          minimum: 1
        ws_url:
          type: string
          description: Relative WS path, e.g. `/api/reports/ws/<task_id>`.
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...